ダウンロードとアップロードとメイン画面をGUIディレクトリに移動
