+++ .gitignore
@@ -1,7 +1,7 @@
 .idea
 test.csv
 Raspberrypi/__pycache__
-Raspberrypi/mainGUI/__pycache__
+Raspberrypi/GUI/__pycache__
 Raspberrypi/*.csv
 Raspberrypi/F2/*.csv
 Raspberrypi/data/
