+++ Raspberrypi/GUI/SplitCharacter.py
