+++ Raspberrypi/GUI/time_compare.py
