+++ Raspberrypi/GUI/widget/syuketu_widget.py
