+++ Raspberrypi/GUI/widget/totalization_widget.py
